move entry to trade specific python files
