+++ basedf.py
@@ -4,18 +4,17 @@ import os
 import pandas as pd
 
 import calculations as calc
-import tradecalcs as tradecalcs
+import settings as s
 
-PATH = './data/'
 
 def prepcoredata():
 
     columns = ['Date', 'Open', 'High', 'Low', 'Close', 'Volume','Adjustment']
     df = pd.DataFrame(columns=columns)
 
-    for csvfile in os.listdir(PATH):
+    for csvfile in os.listdir(s.PATH):
         if csvfile.endswith('.csv'):
-            t = pd.read_csv(PATH+csvfile)
+            t = pd.read_csv(s.PATH+csvfile)
             df = df.append(t, sort=True)
 
     df.drop_duplicates(subset=columns, keep='first', inplace=True)
@@ -33,22 +32,3 @@ def prepcoredata():
     df = calc.priordayvalues(df)
 
     return df
-
-def run8020test(df):
-
-    df['Trigger8020'] = df.apply(lambda x: tradecalcs.trigger8020(x['PriorHigh'],
-        x['PriorLow'],x['PriorOpen'],x['PriorClose']),axis=1)
-
-    df['Trade8020'] = df.apply(lambda x: tradecalcs.trade8020(x['Close'],
-        x['PriorHigh'],x['PriorLow'],x['PriorClose'],x['Trigger8020']),axis=1)
-
-    df.shape
-    df.head(20)
-
-    table = pd.pivot_table(df, index=['Trade8020'], aggfunc='count')
-    print(table['Close'])
-
-
-if __name__ == "__main__":
-    df = prepcoredata()
-    run8020test(df)
