+++ settings.py
@@ -1,0 +1,1 @@
+PATH='./data/msft/'
