+++ trade8020.py
@@ -1,5 +1,22 @@
 import datetime as dt
 import pandas as pd
+import basedf as basedf
+
+
+
+def run8020test(df):
+
+    df['Trigger8020'] = df.apply(lambda x: trigger8020(x['PriorHigh'],
+        x['PriorLow'],x['PriorOpen'],x['PriorClose']),axis=1)
+
+    df['Trade8020'] = df.apply(lambda x: trade8020(x['Close'],
+        x['PriorHigh'],x['PriorLow'],x['PriorClose'],x['Trigger8020']),axis=1)
+
+    df.shape
+    df.head(20)
+
+    table = pd.pivot_table(df, index=['Trade8020'], aggfunc='count')
+    print(table['Close'])
 
 
 def trigger8020(priorhigh, priorlow, prioropen, priorclose):
@@ -36,3 +53,8 @@ def trade8020(close, priorhigh, priorlow, priorclose, trigger8020):
         r = 'notrade'
 
     return r
+
+
+if __name__ == "__main__":
+    df = basedf.prepcoredata()
+    run8020test(df)
